Moved page functionality to new file
